Add ability to limit reporte results to GitHub
When testing the danger-undercover behaviour on Big Pull request we stumped upon some issues all in regards to the size of the reported messages.

The first encountered issue had to deal with the fact that Github API have an abuse mechanism that is triggered by too many in-line messages, hence failing our danger run.
To fix this, we introduced a MAX_INLINE_COMMENTS value which will limit comments to 30, and is configurable.

The second issue had to deal with the size of the single `report` message when the `in-line` option is `false`.
For that, we are trimmering the message to 60k of characters. Given that `danger` does decorate messages further, we have left a good gap to not encounter the size limit reached error. More info about why it is, can be found [here](https://github.com/danger/danger/issues/756)